rag-backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baml-py>=0.87.2",
    "duckdb>=1.2.2",
    "fastembed>=0.6.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "google-genai>=1.14.0",
    "aiohttp>=3.11.18",
    "flasgger>=0.9.7.1",
    "flask[async]>=3.1.0",
    "flask-cors>=5.0.1",
    "gunicorn>=23.0.0",
    "pytest>=8.3.5",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinx-autoapi>=3.6.0",
    "sphinxcontrib-httpdomain>=1.8.1",
    "pytest-mock>=3.14.0",
    "aioresponses>=0.7.8",
    "pytest-asyncio>=0.26.0",
]
